pyproject.toml
[tool.poetry]
name = "goit-web-hw-11"
version = "0.1.0"
description = ""
authors = ["Serhii_Stepanov <stepa.sergey.nov@gmail.com>"]
readme = "README.md"
packages = [{include = "goit_web_hw_11"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.101.0"}
sqlalchemy = "^2.0.19"
psycopg2 = "^2.9.7"
phonenumbers = "^8.13.18"
alembic = "^1.11.2"
fastapi-jwt-auth = "^0.5.0"
uvicorn = "^0.23.2"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
fastapi-mail = "^1.4.1"
python-dotenv = "^1.0.0"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.34.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
